Replace webserver script
1. Bind to 127.0.0.1 instead of 0.0.0.0
2. Switch to python3 http.server instead of ruby WEBrick

Using python3 instead of irb, and binding to 127.0.0.1 instead of
0.0.0.0, improves the speed on macOS of the first page load after
the script is started. Something about irb with WEBrick was
triggering a 5 second delay on the second HTTP GET after the old
webserver was started. It was really odd, and very consistent. May
have had something to do with macOS firewall policy.

Anyhow, using python3 http.server on 127.0.0.1:8000 results in very
fast page loads. Also, doing it this way avoids triggering the OS
security policy prompt that would happen the first time an irb
script tried to listen on 0.0.0.0:8000.