Switch to Cartesian coordinates
Previously, I was using the computer graphics style coordinates
with origin at top-left and +y direction pointing down. But,
mentally keeping track of all the y-axis flips makes it awkward to
work with trig. Also, scaling is a pain with non-centered origin.

This change centers the origin and adds the necessary transforms
to make +y point up.